\name{DavisThin}
\alias{DavisThin}
\docType{data}
\title{Davis's Data on Drive for Thinness}
\usage{DavisThin}
\description{
The \code{DavisThin} data frame has 191 rows and 7 columns.
This is part of a larger dataset for a study of eating disorders.
The 7 variables in the data frame comprise a "drive for thinness"
scale, to be formed by summing the items.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{DT1}{item, coded 0--3}
    \item{DT2}{item, coded 0--3}
    \item{DT3}{item, coded 0--3}
    \item{DT4}{item, coded 0--3}
    \item{DT5}{item, coded 0--3}
    \item{DT6}{item, coded 0--3}
    \item{DT7}{item, coded 0--3}
  }
}

\source{
    Davis, C., G. Claridge, and D. Cerullo (1997) 
    Personality factors predisposing to weight preoccupation: 
    A continuum approach to the association between eating disorders and personality disorders. 
    \emph{Journal of Psychiatric Research} \bold{31}, 467--480. 
    [personal communication from the authors.]
}

\references{
  Fox, J. and Weisberg, S. (2011) 
  \emph{An R Companion to Applied Regression}, Second Edition, Sage.
}


\keyword{datasets}
